pyproject.toml
[tool.poetry]
name = "esa-window-system"
version = "0.1.0"
description = "Software for testing encoding and decoding with the BCJR algorithm"
authors = ["Henri <h.vlot@singlequantum.com>"]
readme = "README.md"
packages = [{include = "esa_window_system"}]

[tool.poetry.dependencies]
python = ">=3.10"
numpy = "^2.0.0"
matplotlib = "^3.7.1"
pandas = "^2.0.2"
Pillow = "^9.5.0"
tqdm = "^4.65.0"
tabulate = "^0.9.0"
scipy = "^1.10.1"
opencv-python = "^4.7.0.72"
openpyxl = "^3.1.2"
cython = "^3.0.11"
numba = "^0.60.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
flake8 = "^6.0.0"
pytest = "^8.1.1"
pytest-benchmark = {extras = ["histogram"], version = "^4.0.0"}
setuptools = "^74.1.2"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
